Fixes to get API extension to work.
Added a function/subroutine to the API to let a programmer set the root
communicator for GA and TCGMSG. At present this breaks the abstractions
at every level in the call stack. The reason is that the call stack is
as follows:

   GA -> ARMCI -> comex -> MPI

   TCGMSG -> MPI

The TCGMSG API does not expose a group or communicator concept. The GA
has processor groups, ARMCI has ARMCI groups, comex has comex groups.
If I want to call MPI_Init at the top, split the MPI_COMM_WORLD communicator
and pass the result sub-communicator all the way down to the MPI layer then
to do that without breaking the abstractions I would have to convert the
MPI sub-communicator into a GA processor group, then into and ARMCI group,
then into a comex group and then back to an MPI communicator. However, until
you reach MPI at the bottom the layers on top have no way of knowning what
the communicator corresponds to. The GA processor group can only be properly
defined when the communicator hits the MPI layer where it can be queried
and the results propagated back up the call stack to properly define
the comex group, the ARMCI group, and the GA processor group.

Implementing these translations properly requires some careful programming
and knowledge of the various layers in the call stack that I don't have.
So this is implementation is nothing more than a hack that allows us to
explore this approach. However, as it stands it is probably unacceptable,
to be incorporated into the main code.